--- a/core/base/discreteGradient/DiscreteGradient.cpp
+++ b/core/base/discreteGradient/DiscreteGradient.cpp
@@ -191,21 +191,19 @@ bool DiscreteGradient::isCellCritical(const Cell &cell) const {
 }
 
 int DiscreteGradient::setManifoldSize(
-  const size_t nCritPoints,
-  const std::vector<size_t> &nCriticalPointsByDim,
+  const std::array<std::vector<SimplexId>, 4> &criticalCellsByDim,
   const SimplexId *const ascendingManifold,
   const SimplexId *const descendingManifold,
   std::vector<SimplexId> &manifoldSize) const {
 
-  // in the manifoldSize vector, critical points are sorted first by
-  // dim then by index
+  const auto nCritPoints{
+    criticalCellsByDim[0].size() + criticalCellsByDim[1].size()
+    + criticalCellsByDim[2].size() + criticalCellsByDim[3].size()};
 
-  // ascendingManifold (resp. descendingManifold) region indices are
-  // numbered from 0 to #maxima == nCriticalPointsByDim.back()
-  // (resp. #minina == nCriticalPointsByDim[0])
+  const auto dim{this->dimensionality_};
 
   if(nCritPoints == 0
-     || (nCriticalPointsByDim[0] == 0 && nCriticalPointsByDim.back() == 0)) {
+     || (criticalCellsByDim[0].empty() && criticalCellsByDim[dim].empty())) {
     // no critical points || no extrema
     return 0;
   }
@@ -213,7 +211,7 @@ int DiscreteGradient::setManifoldSize(
   manifoldSize.resize(nCritPoints, 0);
 
   // descending manifold cells size
-  if(nCriticalPointsByDim[0] > 0) {
+  if(!criticalCellsByDim[0].empty()) {
     for(SimplexId i = 0; i < numberOfVertices_; ++i) {
       if(descendingManifold[i] != -1) {
         manifoldSize[descendingManifold[i]]++;
@@ -221,9 +219,9 @@ int DiscreteGradient::setManifoldSize(
     }
   }
 
-  if(nCriticalPointsByDim.back() > 0) {
+  if(!criticalCellsByDim[dim].empty()) {
     // index of first maximum in critical points array
-    const auto nFirstMaximum{nCritPoints - nCriticalPointsByDim.back()};
+    const auto nFirstMaximum{nCritPoints - criticalCellsByDim[dim].size()};
 
     // ascending manifold cells size
     for(SimplexId i = 0; i < numberOfVertices_; ++i) {